simplification from the Protect version to basic Galileo logging
