+++ python/chatbot/elevenlabs-chatbot/.env.example
@@ -4,9 +4,6 @@ ELEVENLABS_AGENT_ID=your-agent-id
 
 # Galileo Configuration (for local instance)
 GALILEO_API_KEY=your-galileo-api-key
-GALILEO_CONSOLE_URL=
+GALILEO_CONSOLE_URL=your-galileo-console-url
 GALILEO_PROJECT_NAME=elevenlabs-voice-poc
-GALILEO_LOG_STREAM=voice-conversations
-
-GALILEO_PROTECT_ENABLED=true
-GALILEO_PROTECT_STAGE_ID=your-stage-id
+GALILEO_LOG_STREAM=voice-chatbot
