+++ python/chatbot/elevenlabs-chatbot/requirements.txt
@@ -1,6 +1,4 @@
-websockets>=12.0
 python-dotenv>=1.0.0
-pydantic-settings>=2.0.0
 elevenlabs>=1.0.0
 galileo>=1.34.0
 pyaudio>=0.2.14
